feat: 部分模块化工作处理
